[Optimizer Overlap] Move hooks to own file
Moves current prototype optimizer overlap to its own file for a better
namespace. No code changes besides a few comment fixes. Note that this code is
still prototype and not expected to be used by an end user.

Differential Revision: [D33662678](https://our.internmc.facebook.com/intern/diff/D33662678/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D33662678/)!

[ghstack-poisoned]
